+++ src/Language/AST/TypeSystemExtensionNode.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace GraphQL\Language\AST;
+
+/**
+ * export type TypeSystemExtensionNode = SchemaExtensionNode | TypeExtensionNode;
+ */
+interface TypeSystemExtensionNode extends DefinitionNode
+{
+}